res fixed the csNullCacheManager::CacheData() implementation   being different from prototype.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@18370 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/nulcache.cpp
@@ -48,7 +48,7 @@ void csNullCacheManager::SetCurrentScope (const char*)
 {
 }
 
-bool csNullCacheManager::CacheData (void*, uint32,
+bool csNullCacheManager::CacheData (void*, size_t,
   	const char*, const char*, uint32)
 {
   return false;
